requirements.txt
streamlit>=1.32
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
torch
torchvision
requests>=2.31.0 